Refactor Role class into own file
